Move pkg/util/netutils to mostly pkg/network/common
