+++ pkg/network/common/common.go
@@ -11,8 +11,8 @@ import (
 
 	networkapi "github.com/openshift/api/network/v1"
 	networkclient "github.com/openshift/client-go/network/clientset/versioned"
+	"github.com/openshift/library-go/pkg/network/networkutils"
 	configapi "github.com/openshift/origin/pkg/cmd/server/apis/config"
-	"github.com/openshift/origin/pkg/util/netutils"
 )
 
 func HostSubnetToString(subnet *networkapi.HostSubnet) string {
@@ -47,7 +47,7 @@ func ParseNetworkInfo(clusterNetwork []networkapi.ClusterNetworkEntry, serviceNe
 	var cns []ClusterNetwork
 
 	for _, entry := range clusterNetwork {
-		cidr, err := netutils.ParseCIDRMask(entry.CIDR)
+		cidr, err := networkutils.ParseCIDRMask(entry.CIDR)
 		if err != nil {
 			_, cidr, err = net.ParseCIDR(entry.CIDR)
 			if err != nil {
@@ -58,7 +58,7 @@ func ParseNetworkInfo(clusterNetwork []networkapi.ClusterNetworkEntry, serviceNe
 		cns = append(cns, ClusterNetwork{ClusterCIDR: cidr, HostSubnetLength: entry.HostSubnetLength})
 	}
 
-	sn, err := netutils.ParseCIDRMask(serviceNetwork)
+	sn, err := networkutils.ParseCIDRMask(serviceNetwork)
 	if err != nil {
 		_, sn, err = net.ParseCIDR(serviceNetwork)
 		if err != nil {
@@ -167,4 +167,53 @@ func GetNetworkInfo(networkClient networkclient.Interface) (*NetworkInfo, error)
 	}
 
 	return ParseNetworkInfo(cn.ClusterNetworks, cn.ServiceNetwork, cn.VXLANPort)
+}
+
+// Generate the default gateway IP Address for a subnet
+func GenerateDefaultGateway(sna *net.IPNet) net.IP {
+	ip := sna.IP.To4()
+	return net.IPv4(ip[0], ip[1], ip[2], ip[3]|0x1)
+}
+
+// Return Host IP Networks
+// Ignores provided interfaces and filters loopback and non IPv4 addrs.
+func GetHostIPNetworks(skipInterfaces []string) ([]*net.IPNet, []net.IP, error) {
+	hostInterfaces, err := net.Interfaces()
+	if err != nil {
+		return nil, nil, err
+	}
+
+	skipInterfaceMap := make(map[string]bool)
+	for _, ifaceName := range skipInterfaces {
+		skipInterfaceMap[ifaceName] = true
+	}
+
+	errList := []error{}
+	var hostIPNets []*net.IPNet
+	var hostIPs []net.IP
+	for _, iface := range hostInterfaces {
+		if skipInterfaceMap[iface.Name] {
+			continue
+		}
+
+		ifAddrs, err := iface.Addrs()
+		if err != nil {
+			errList = append(errList, err)
+			continue
+		}
+		for _, addr := range ifAddrs {
+			ip, ipNet, err := net.ParseCIDR(addr.String())
+			if err != nil {
+				errList = append(errList, err)
+				continue
+			}
+
+			// Skip loopback and non IPv4 addrs
+			if !ip.IsLoopback() && ip.To4() != nil {
+				hostIPNets = append(hostIPNets, ipNet)
+				hostIPs = append(hostIPs, ip)
+			}
+		}
+	}
+	return hostIPNets, hostIPs, kerrors.NewAggregate(errList)
 }
