--- pkg/util/netutils/README.md
@@ -1,2 +1,0 @@
-# netutils
-Miscellaneous network utilities
